add guest page

+++ src/app/shared/vouchers service/vouchers-service.service.ts
@@ -1,12 +1,17 @@
 import { HttpClient } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { Injectable, OnDestroy } from '@angular/core';
+import { Subscription } from 'rxjs';
 import { Guest, Voucher } from '../models';
 
 @Injectable({
   providedIn: 'root',
 })
-export class VouchersServiceService {
+export class VouchersServiceService implements OnDestroy {
+  useVoucherSub$$!: Subscription;
+  useVoucherPutSub$$!: Subscription;
+
   url = 'https://airbus-900f9-default-rtdb.firebaseio.com/guests.json';
+
   constructor(private http: HttpClient) {}
 
   IdGenerator() {
@@ -34,32 +39,31 @@ export class VouchersServiceService {
   }
 
   useVoucher(voucherId: string) {
-    this.http.get(this.url).subscribe((data) => {
+    this.useVoucherSub$$ = this.http.get(this.url).subscribe((data) => {
       const guestsArray = Object.values(data);
-      this.http
-        .put(this.url, this.deleteVoucher(guestsArray, voucherId))
-        .subscribe((response) => {
-          console.log('response :', response);
-        });
+      let indexOfGuest = this.indexOfGuestFromId(
+        guestsArray,
+        voucherId.slice(0, 13)
+      );
+      if (
+        this.deleteVoucher(guestsArray, voucherId)[indexOfGuest].vouchersLis
+          .length !== 0
+      ) {
+        this.useVoucherPutSub$$ = this.http
+          .put(this.url, this.deleteVoucher(guestsArray, voucherId))
+          .subscribe();
+      }
     });
   }
 
   deleteVoucher(arrayOfGuests: Guest[], voucherId: string) {
     let updatedArray: Guest[] = [];
     let guestId = voucherId.slice(0, 13);
-    console.log(guestId);
     for (let i = 0; i < arrayOfGuests.length; i++) {
       if (arrayOfGuests[i].id === guestId) {
-        console.log('foundGuestId', arrayOfGuests[i].vouchersLis);
-
-        console.log(voucherId);
         let newVouchList = this.filterVouchers(
           arrayOfGuests[i].vouchersLis,
           voucherId
-        );
-        console.log(
-          'filteredArr',
-          this.filterVouchers(arrayOfGuests[i].vouchersLis, voucherId)
         );
         updatedArray.push(
           new Guest(
@@ -67,7 +71,6 @@ export class VouchersServiceService {
             arrayOfGuests[i].name,
             arrayOfGuests[i].roomNumber,
             arrayOfGuests[i].type,
-            arrayOfGuests[i].vouchers,
             arrayOfGuests[i].validUntill,
             newVouchList,
             arrayOfGuests[i].createdDate
@@ -80,13 +83,34 @@ export class VouchersServiceService {
     return updatedArray;
   }
 
-  filterVouchers(array: Voucher[], id: string) {
+  filterVouchers(filteringArray: Voucher[], id: string) {
     let filteredArray = [];
-    for (let i = 0; i < array.length; i++) {
-      if (array[i].id !== id) {
-        filteredArray.push(array[i]);
+    let voucherNotFound = false;
+    for (let i = 0; i < filteringArray.length; i++) {
+      if (filteringArray[i].id !== id) {
+        filteredArray.push(filteringArray[i]);
       }
     }
-    return filteredArray;
+    if (filteringArray.length === filteredArray.length) {
+      console.log('voucher doesnt exist, error');
+      return [];
+    } else {
+      return filteredArray;
+    }
+  }
+
+  indexOfGuestFromId(array: Guest[], id: string) {
+    let index = -1;
+    for (let i = 0; i < array.length; i++) {
+      if (array[i].id === id) {
+        index = i;
+      }
+    }
+    return index;
+  }
+
+  ngOnDestroy(): void {
+    this.useVoucherSub$$.unsubscribe();
+    this.useVoucherPutSub$$.unsubscribe();
   }
 }
